Expose DialogueParser and DialogueSerializer in the public API
Both almond-cloud and thingpedia-common-devices will make use
of these.
