+++ lib/index.js
@@ -25,7 +25,12 @@ const { ParaphraseValidator, ParaphraseValidatorFilter } = require('./dataset-to
 const ValidationHITCreator = require('./dataset-tools/mturk/paraphrase-validation');
 
 const DatasetAugmenter = require('./dataset-tools/augmentation');
-const { DatasetParser, DatasetStringifier } = require('./dataset-tools/parsers');
+const {
+    DatasetParser,
+    DatasetStringifier,
+    DialogueParser,
+    DialogueSerializer
+} = require('./dataset-tools/parsers');
 const DatasetSplitter = require('./dataset-tools/splitter');
 const {
     SentenceEvaluatorStream,
@@ -80,6 +85,8 @@ module.exports = {
     DatasetStringifier,
     DatasetAugmenter,
     DatasetSplitter,
+    DialogueParser,
+    DialogueSerializer,
 
     I18n,
     MTurk,
